PaaS/Development/Security/requirements.txt
boto3>=1.26.0
paramiko>=3.0.0
ansible>=6.0.0
python-nmap>=0.7.1
requests>=2.28.0
pywinrm>=0.4.3
cryptography>=39.0.0
pyyaml>=6.0.0
python-dotenv>=0.21.0
schedule>=1.1.0
